# src/main/resources/Swagger-UI/swagger.json
{
   "swagger":"2.0",
   "info":{
      "version":"Swagger Server",
      "title":""
   },
   "host":"localhost:8080",
   "basePath":"/jesuswalkrest",
...
   "schemes":[
      "http"
   ],
   "paths":{
      "/counselors":{
         "get":{
            "tags":[
               "counselors"
            ],
            "operationId":"retrieveCounselors",
            "parameters":[

            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         },
         "post":{
            "tags":[
               "counselors"
            ],
            "operationId":"createCounselor",
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "required":false,
                  "schema":{
                     "$ref":"#/definitions/CounselorEntity"
                  }
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         },
         "put":{
            "tags":[
               "counselors"
            ],
            "operationId":"updateCounselor",
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "required":false,
                  "schema":{
                     "$ref":"#/definitions/CounselorEntity"
                  }
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         },
         "delete":{
            "tags":[
               "counselors"
            ],
            "operationId":"deleteCounselor",
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "required":false,
                  "schema":{
                     "$ref":"#/definitions/CounselorEntity"
                  }
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         }
      },
...
      "/staffs":{
         "get":{
            "tags":[
               "staffs"
            ],
            "operationId":"retrieveStaffs",
            "parameters":[

            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         },
         "post":{
            "tags":[
               "staffs"
            ],
            "operationId":"createStaff",
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "required":false,
                  "schema":{
                     "$ref":"#/definitions/StaffEntity"
                  }
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         },
         "put":{
            "tags":[
               "staffs"
            ],
            "operationId":"updateStaff",
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "required":false,
                  "schema":{
                     "$ref":"#/definitions/StaffEntity"
                  }
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         },
         "delete":{
            "tags":[
               "staffs"
            ],
            "operationId":"deleteStaff",
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "required":false,
                  "schema":{
                     "$ref":"#/definitions/StaffEntity"
                  }
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         }
      },
...
      "/students":{
         "get":{
            "tags":[
               "students"
            ],
            "operationId":"retrieveStudents",
            "parameters":[

            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         },
         "post":{
            "tags":[
               "students"
            ],
            "operationId":"createStudent",
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "required":false,
                  "schema":{
                     "$ref":"#/definitions/StudentEntity"
                  }
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         },
         "put":{
            "tags":[
               "students"
            ],
            "operationId":"updateStudent",
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "required":false,
                  "schema":{
                     "$ref":"#/definitions/StudentEntity"
                  }
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         },
         "delete":{
            "tags":[
               "students"
            ],
            "operationId":"deleteStudent",
            "parameters":[
               {
                  "in":"body",
                  "name":"body",
                  "required":false,
                  "schema":{
                     "$ref":"#/definitions/StudentEntity"
                  }
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         }
      },
      "/students/age/{age}":{
         "get":{
            "tags":[
               "students"
            ],
            "operationId":"retrieveStudentByAge",
            "parameters":[
               {
                  "name":"age",
                  "in":"path",
                  "required":true,
                  "type":"string"
               }
            ],
            "responses":{
               "default":{
                  "description":"successful operation"
               }
            }
         }
      },
...
   },
   "definitions":{
...
      "ChurchEntity":{
         "type":"object",
         "properties":{
            "id":{
               "type":"string"
            },
            "creationDate":{
               "type":"string",
               "format":"date-time"
            },
            "lastUpdated":{
               "type":"string",
               "format":"date-time"
            },
            "name":{
               "type":"string"
            },
            "seniorPastorFirstName":{
               "type":"string"
            },
            "seniorPastorLastName":{
               "type":"string"
            },
            "youthPastorFirstName":{
               "type":"string"
            },
            "youthPastorLastName":{
               "type":"string"
            },
            "address":{
               "$ref":"#/definitions/Address"
            }
         }
      },
      "Address":{
         "type":"object",
         "properties":{
            "street":{
               "type":"string"
            },
            "city":{
               "type":"string"
            },
            "state":{
               "type":"string"
            },
            "zip":{
               "type":"string"
            }
         }
      },
      "Mother":{
         "type":"object",
         "properties":{
            "firstName":{
               "type":"string"
            },
            "lastName":{
               "type":"string"
            }
         }
      },
      "CounselorEntity":{
         "type":"object",
         "properties":{
            "id":{
               "type":"string"
            },
            "creationDate":{
               "type":"string",
               "format":"date-time"
            },
            "lastUpdated":{
               "type":"string",
               "format":"date-time"
            },
            "firstName":{
               "type":"string"
            },
            "lastName":{
               "type":"string"
            },
            "age":{
               "type":"integer",
               "format":"int32"
            },
            "birthdate":{
               "type":"string",
               "format":"date-time"
            },
            "sex":{
               "type":"string"
            },
            "registerDate":{
               "type":"string",
               "format":"date-time"
            },
            "shirtSize":{
               "type":"string"
            },
            "dormNumber":{
               "type":"string"
            },
            "dormKey":{
               "type":"string"
            },
            "profile":{
               "$ref":"#/definitions/Profile"
            },
            "address":{
               "$ref":"#/definitions/Address"
            },
            "church":{
               "$ref":"#/definitions/ChurchEntity"
            },
            "groupName":{
               "type":"string"
            },
            "groupGrade":{
               "type":"string"
            },
            "students":{
               "type":"array",
               "uniqueItems":true,
               "items":{
                  "$ref": "StudentEntity"
               }
            }
         }
      },
      "Profile":{
         "type":"object",
         "properties":{
            "homeNumber":{
               "type":"string"
            },
            "cellNumber":{
               "type":"string"
            },
            "instagram":{
               "type":"string"
            },
            "snapchat":{
               "type":"string"
            },
            "twitter":{
               "type":"string"
            }
         }
      },
      "StudentEntity":{
         "type":"object",
         "properties":{
            "id":{
               "type":"string"
            },
            "creationDate":{
               "type":"string",
               "format":"date-time"
            },
            "lastUpdated":{
               "type":"string",
               "format":"date-time"
            },
            "firstName":{
               "type":"string"
            },
            "lastName":{
               "type":"string"
            },
            "age":{
               "type":"integer",
               "format":"int32"
            },
            "birthdate":{
               "type":"string",
               "format":"date-time"
            },
            "sex":{
               "type":"string"
            },
            "registerDate":{
               "type":"string",
               "format":"date-time"
            },
            "shirtSize":{
               "type":"string"
            },
            "dormNumber":{
               "type":"string"
            },
            "dormKey":{
               "type":"string"
            },
            "profile":{
               "$ref":"#/definitions/Profile"
            },
            "address":{
               "$ref":"#/definitions/Address"
            },
            "church":{
               "$ref":"#/definitions/ChurchEntity"
            },
            "groupName":{
               "type":"string"
            },
            "grade":{
               "type":"string"
            },
            "mother":{
               "$ref":"#/definitions/Mother"
            },
            "father":{
               "$ref":"#/definitions/Father"
            },
            "counselors":{
               "type":"array",
               "uniqueItems":true,
               "items":{
                  "$ref":"#/definitions/CounselorEntity"
               }
            }
         }
      },
      "Father":{
         "type":"object",
         "properties":{
            "firstName":{
               "type":"string"
            },
            "lastName":{
               "type":"string"
            }
         }
      },
      "StaffEntity":{
         "type":"object",
         "properties":{
            "id":{
               "type":"string"
            },
            "creationDate":{
               "type":"string",
               "format":"date-time"
            },
            "lastUpdated":{
               "type":"string",
               "format":"date-time"
            },
            "firstName":{
               "type":"string"
            },
            "lastName":{
               "type":"string"
            },
            "age":{
               "type":"integer",
               "format":"int32"
            },
            "birthdate":{
               "type":"string",
               "format":"date-time"
            },
            "sex":{
               "type":"string"
            },
            "registerDate":{
               "type":"string",
               "format":"date-time"
            },
            "shirtSize":{
               "type":"string"
            },
            "dormNumber":{
               "type":"string"
            },
            "dormKey":{
               "type":"string"
            },
            "profile":{
               "$ref":"#/definitions/Profile"
            },
            "address":{
               "$ref":"#/definitions/Address"
            },
            "church":{
               "$ref":"#/definitions/ChurchEntity"
            },
            "counseling":{
               "type":"boolean",
               "default":false
            }
         }
      }
   }
}